Reestructuracion del codigo.
